Improve declarations, fix comment, and move thread check
